+++ replication/report/replication_report.go
@@ -97,18 +97,9 @@ type StepReport struct {
 	Info *StepInfo
 }
 
-type EncryptedEnum string
-
-const (
-	EncryptedTrue            EncryptedEnum = "yes"
-	EncryptedFalse           EncryptedEnum = "no"
-	EncryptedSenderDependent EncryptedEnum = "sender-dependent"
-)
-
 type StepInfo struct {
 	From, To        string
 	Resumed         bool
-	Encrypted       EncryptedEnum
 	BytesExpected   uint64
 	BytesReplicated uint64
 }